+++ src/duk_js_bytecode.h
@@ -188,8 +188,7 @@ typedef duk_uint32_t duk_instr_t;
 #define DUK_BC_TRYCATCH_FLAG_WITH_BINDING   (1 << 3)
 
 /* DUK_OP_RETURN flags in A */
-#define DUK_BC_RETURN_FLAG_FAST             (1 << 0)
-#define DUK_BC_RETURN_FLAG_HAVE_RETVAL      (1 << 1)
+#define DUK_BC_RETURN_FLAG_HAVE_RETVAL      (1 << 0)
 
 /* DUK_OP_DECLVAR flags in A; bottom bits are reserved for propdesc flags (DUK_PROPDESC_FLAG_XXX) */
 #define DUK_BC_DECLVAR_FLAG_UNDEF_VALUE     (1 << 4)  /* use 'undefined' for value automatically */