+++ tests/perf/test-call-native.js
@@ -1,0 +1,19 @@
+function test() {
+    var arr = [];
+    for (var i = 0; i < 1e3; i++) {
+        arr[i] = i;
+    }
+    for (var i = 0; i < 3e4; i++) {
+        // Each .forEach() call is a native call so that return 123 causes a
+        // return-to-native situation which was previously a longjmp.
+
+        arr.forEach(function () { return 123; });
+    }
+}
+
+try {
+    test();
+} catch (e) {
+    print(e.stack || e);
+    throw e;
+}